Medical Chatbot/requirements.txt
streamlit
google-generativeai
python-dotenv
transformers
sentencepiece
# llama==0.0.1